# sandbox/api-go/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/beer": {
            "post": {
                "description": "Insert single beer to beer_log with POST body and no URL params",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "text/plain"
                ],
                "tags": [
                    "beer"
                ],
                "summary": "Create single record of pour from provided JSON",
                "parameters": [
                    {
                        "description": "Beer to be recorded",
                        "name": "beer",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.Beer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/beer/{user}": {
            "put": {
                "description": "Update single (last poured) beer to beer_log with PUT request (no body) and the consumer as a URL parameter",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "text/plain"
                ],
                "tags": [
                    "beer"
                ],
                "summary": "Create single record of pour from provided JSON",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User to attribute the last pour to",
                        "name": "user",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/kegs/all": {
            "get": {
                "description": "Gets all kegs stored in DB, represented in JSON format",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "kegs"
                ],
                "summary": "Get all kegs",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/types.Keg"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "types.Beer": {
            "type": "object",
            "properties": {
                "beer_id": {
                    "type": "string"
                },
                "beer_name": {
                    "type": "string"
                },
                "consumer": {
                    "type": "string"
                },
                "oz_poured": {
                    "type": "number"
                },
                "tap_id": {
                    "type": "integer"
                },
                "time": {
                    "type": "string"
                }
            }
        },
...
        "types.Keg": {
            "type": "object",
            "properties": {
                "abv": {
                    "type": "number"
                },
                "capacity": {
                    "type": "integer"
                },
                "date_kicked": {
                    "type": "string"
                },
                "date_tapped": {
                    "type": "string"
                },
                "days_to_consume": {
                    "type": "integer"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "remaining": {
                    "type": "number"
                },
                "status": {
                    "type": "integer"
                },
                "tap": {
                    "type": "integer"
                }
            }
        },
...
    }
}